Condense repeated panel info into Panel component and create DeletePanel
