--- lang2views-frontend/src/Components/VideoList/Organize/Organize.css
@@ -1,86 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-.organize-overlay {
-    position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    background-color: rgba(0, 0, 0, 0.5);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.organize-content {
-    background: white;
-    padding: 20px;
-    border-radius: 5px;
-    width: 500px;
-    text-align: center;
-}
-
-.video-name-in-organize {
-    color: #8A8A8A;
-    font-weight: bold;
-}
-
-.text-with-input-information {
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-start;
-    align-items: center;
-    gap: 20px;
-    margin-top: 20px;
-}
-
-.widget-text {
-    font-size: 16px;
-    margin: 0;
-    font-weight: bold;
-    width: 155px;
-}
-
-.widget-input {
-    background-color: white;
-    border-radius: 5px;
-    width: 225px;
-    height: 25px;
-    color: black;
-    padding-left: 10px;
-}
-
-.arrow-diagonal {
-    height: 22px;
-    width: 22px;
-    margin-left: -42px;
-    border-radius: 3px;
-    border: none;
-}
-
-.organize-button {
-    background-color: #4285F4;
-    color: white;
-    font-size: 18px;
-    border-radius: 10px;
-    border: none;
-    width: 125px;
-    height: 40px;
-    margin-top: 20px;
-}
-
-.go-back-button {
-    background-color: #4285F4;
-    color: white;
-    font-size: 18px;
-    border-radius: 10px;
-    border: none;
-    width: 125px;
-    height: 40px;
-    margin-top: 20px;
-    margin-left: 10px;
-}
